FOGL-4776 Fix RPi compile issue

+++ C/plugins/storage/sqlite/common/connection.cpp
@@ -1173,7 +1173,7 @@ std::size_t arr = data.find("inserts");
 			else if (itr->value.IsDouble())
 				values.append(itr->value.GetDouble());
 			else if (itr->value.IsInt64())
-				values.append(itr->value.GetInt64());
+				values.append((long)itr->value.GetInt64());
 			else if (itr->value.IsInt())
 				values.append(itr->value.GetInt());
 			else if (itr->value.IsObject())
@@ -1345,7 +1345,7 @@ SQLBuffer	sql;
 					else if (itr->value.IsDouble())
 						sql.append(itr->value.GetDouble());
 					else if (itr->value.IsInt64())
-						sql.append(itr->value.GetInt64());
+						sql.append((long)itr->value.GetInt64());
 					else if (itr->value.IsNumber())
 						sql.append(itr->value.GetInt());
 					else if (itr->value.IsObject())
@@ -1423,7 +1423,7 @@ SQLBuffer	sql;
 					else if (value.IsDouble())
 						sql.append(value.GetDouble());
 					else if (value.IsInt64())
-						sql.append(value.GetInt64());
+						sql.append((long)value.GetInt64());
 					else if (value.IsNumber())
 						sql.append(value.GetInt());
 					else if (value.IsObject())
@@ -1535,7 +1535,7 @@ SQLBuffer	sql;
 					}
 					else if (value.IsInt64())
 					{
-						sql.append(value.GetInt64());
+						sql.append((long)value.GetInt64());
 					}
 					else if (value.IsNumber())
 					{
